requirements.txt
apache-flink==1.17.2-scala_2.12

boto3==1.34.98
pandas==1.3.5
pyyaml